requirements.txt
django==4.2.7 
djangorestframework==3.14.0 
pyodbc>= 5.1
